+++ builders/JobBuilder.js
@@ -1,0 +1,33 @@
+import Luxon from "@bejibun/utils/facades/Luxon";
+import JobModel from "../models/JobModel";
+export default class JobBuilder {
+    queue;
+    now;
+    availableAt;
+    args;
+    constructor() {
+        this.now = Luxon.DateTime.now().toUnixInteger();
+        this.availableAt = this.now;
+        this.args = [];
+    }
+    setQueue(queue) {
+        this.queue = queue;
+        return this;
+    }
+    dispatch(...args) {
+        this.args.push(...args);
+        return this;
+    }
+    delay(delay) {
+        this.availableAt = this.now + delay;
+        return this;
+    }
+    async send() {
+        await JobModel.create({
+            queue: this.queue,
+            payload: JSON.stringify(this.args),
+            available_at: this.availableAt,
+            created_at: this.now
+        });
+    }
+}