+++ loader/NamespaceLoader.js
@@ -1,0 +1,6 @@
+import NamespaceBuilder from "../builders/NamespaceBuilder";
+export default class NamespaceLoader {
+    static async load(directory) {
+        return await new NamespaceBuilder().load(directory);
+    }
+}