+++ models/JobModel.d.ts
@@ -1,0 +1,18 @@
+import { ModelOptions, QueryBuilder, QueryContext } from "objection";
+declare const JobModel_base: any;
+export default class JobModel extends JobModel_base {
+    static tableName: string;
+    static idColumn: string;
+    static updatedColumn: null;
+    static deletedColumn: null;
+    static QueryBuilder: typeof QueryBuilder;
+    $beforeUpdate(opt: ModelOptions, queryContext: QueryContext): void;
+    id: bigint;
+    queue: string;
+    payload: string;
+    attempts: bigint;
+    reserved_at: bigint | null;
+    available_at: bigint;
+    created_at: bigint;
+}
+export {};