+++ src/bases/BaseModel.ts
@@ -1,7 +1,5 @@
-import App from "@bejibun/app";
-import {defineValue, isEmpty} from "@bejibun/utils";
-import Str from "@bejibun/utils/facades/Str";
-import {DateTime} from "luxon";
+import {defineValue, isEmpty, isNotEmpty} from "@bejibun/utils";
+import Luxon from "@bejibun/utils/facades/Luxon";
 import {
     Constructor,
     Model,
@@ -12,17 +10,12 @@ import {
     QueryContext,
     TransactionOrKnex
 } from "objection";
-import {relative, sep} from "path";
-import {fileURLToPath} from "url";
 import ModelNotFoundException from "@/exceptions/ModelNotFoundException";
 import SoftDeletes from "@/facades/SoftDeletes";
+import RuntimeException from "@/exceptions/RuntimeException";
 
-export interface BaseColumns {
-    id: bigint | number;
-    created_at: DateTime | string;
-    updated_at: DateTime | string;
-    deleted_at: DateTime | string | null;
-}
+export type Timestamp = typeof Luxon.DateTime | Date | string;
+export type NullableTimestamp = Timestamp | null;
 
 class BunQueryBuilder<M extends Model, R = M[]> extends SoftDeletes<M, R> {
     // @ts-ignore
@@ -40,37 +33,43 @@ class BunQueryBuilder<M extends Model, R = M[]> extends SoftDeletes<M, R> {
 }
 
 // @ts-ignore
-export default class BaseModel extends Model implements BaseColumns {
+export default class BaseModel extends Model {
+    protected static _namespace: string;
+
     public static tableName: string;
     public static idColumn: string;
+    public static createdColumn: string = "created_at";
+    public static updatedColumn: string = "updated_at";
     public static deletedColumn: string = "deleted_at";
 
     public static QueryBuilder = BunQueryBuilder;
 
     declare id: number | bigint;
-    declare created_at: DateTime | string;
-    declare updated_at: DateTime | string;
-    declare deleted_at: DateTime | string | null;
 
     public static get namespace(): string {
-        const filePath = fileURLToPath(import.meta.url);
-        const rel = relative(App.Path.rootPath(), filePath);
-        const withoutExt = rel.replace(/\.[tj]s$/, "");
-        const namespaces = withoutExt.split(sep);
-        namespaces.pop();
-        namespaces.push(this.name);
+        if (isEmpty(this._namespace)) throw new RuntimeException(`Model namespace not registered for [${this.name}]`);
 
-        return namespaces.map(part => Str.toPascalCase(part)).join("/");
+        return this._namespace;
     }
 
     $beforeInsert(queryContext: QueryContext): void {
-        const now = DateTime.now();
-        this.created_at = now;
-        this.updated_at = now;
+        const now = Luxon.DateTime.now() as any;
+        if (isNotEmpty((this as any)[(this.constructor as any).createdColumn])) {
+            (this as any)[(this.constructor as any).createdColumn] = now;
+        }
+        if (isNotEmpty((this as any)[(this.constructor as any).updatedColumn])) {
+            (this as any)[(this.constructor as any).updatedColumn] = now;
+        }
     }
 
     $beforeUpdate(opt: ModelOptions, queryContext: QueryContext): void {
-        this.updated_at = DateTime.now();
+        if (isNotEmpty((this as any)[(this.constructor as any).updatedColumn])) {
+            (this as any)[(this.constructor as any).updatedColumn] = Luxon.DateTime.now() as any;
+        }
+    }
+
+    public static setNamespace(namespace: string): void {
+        this._namespace = namespace;
     }
 
     public static query<T extends Model>(this: Constructor<T>, trxOrKnex?: TransactionOrKnex): QueryBuilderType<T> {
@@ -98,7 +97,7 @@ export default class BaseModel extends Model implements BaseColumns {
     }
 
     public static async findOrFail<T extends Model>(this: T, id: bigint | number | string): Promise<T> {
-        const result = await (this as any).find(id);
+        const result: T = await (this as any).find(id);
 
         if (isEmpty(result)) throw new ModelNotFoundException(`No query results for model [${(this as any).namespace}] [${id}].`);