+++ src/builders/RouterBuilder.ts
@@ -14,7 +14,6 @@ import Logger from "@bejibun/logger";
 import {defineValue, isEmpty, isModuleExists, isNotEmpty} from "@bejibun/utils";
 import HttpMethodEnum from "@bejibun/utils/enums/HttpMethodEnum";
 import Enum from "@bejibun/utils/facades/Enum";
-import path from "path";
 import BaseController from "@/bases/BaseController";
 import RouterException from "@/exceptions/RouterException";
 
@@ -24,9 +23,9 @@ export interface ResourceOptions {
 }
 
 export default class RouterBuilder {
-    private basePath: string = "";
-    private middlewares: Array<IMiddleware> = [];
-    private baseNamespace: string = "app/controllers";
+    protected basePath: string = "";
+    protected middlewares: Array<IMiddleware> = [];
+    protected baseNamespace: string = "app/controllers";
 
     public prefix(basePath: string): RouterBuilder {
         this.basePath = basePath;
@@ -354,7 +353,7 @@ export default class RouterBuilder {
             throw new RouterException(`Invalid router controller definition: ${definition}.`);
         }
 
-        const controllerPath = path.resolve(App.Path.rootPath(), defineValue(overrideNamespace, this.baseNamespace));
+        const controllerPath = App.Path.rootPath(defineValue(overrideNamespace, this.baseNamespace));
         let location: any = null;
 
         try {