+++ stubs/jobs/TemplateJob.ts
@@ -1,0 +1,12 @@
+import BaseJob from "@bejibun/core/bases/BaseJob";
+
+export default class TemplateJob extends BaseJob {
+    /**
+     * Execute the job.
+     *
+     * @var $arguments Array<any>
+     */
+    public async handle(args: Array<any>): Promise<void> {
+        //
+    }
+}